third_party/WebKit/LayoutTests/html/dialog/submit-dialog-close-event-expected.txt

Tests submitting a dialog on a close event triggered by a previous submission.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dialog.open is false
PASS dialog.returnValue is "Goodbye"
PASS dialog.open is false
PASS dialog.returnValue is "Hello"
PASS successfullyParsed is true

TEST COMPLETE
